Move transpiler macro-expansion into dedicated macros/ domain
Original location: src/transpiler/src/macro-expansion.ts (741 lines, flat file
directly under src/).

New location: src/transpiler/src/macros/macro-expansion.ts, accompanied by a
new src/transpiler/src/macros/index.ts barrel.

Rationale: The transpiler workspace organizes its sibling domains (api/,
emitter/, event-context/) into named subdirectories, each fronted by an
index.ts barrel that consumers import via "./<domain>/index.js". The
macro-expansion module was the lone outlier: a single 741-line file living at
the top of src/, imported by src/index.ts directly as "./macro-expansion.js"
instead of through a barrel. This made the workspace's import graph
inconsistent and forced every consumer to know the implementation file by
name rather than the domain it belonged to.

Promoting the module into a macros/ directory establishes a clear boundary
matching the existing api/, emitter/, and event-context/ siblings. The new
macros/index.ts re-exports the same public surface, so external consumers
that already used `Transpiler.expandProjectMacros` and friends continue to
work unchanged. Internal imports in src/index.ts and api/gml-transpiler.ts
now route through the barrel or the file path under macros/, matching the
pattern used by every other domain in the workspace.

Verified: pnpm -w run build:ts passes; all 685 transpiler tests pass
(including the dedicated macro-expansion-source, macro-transpile, and index
suites); pnpm -w run lint:quiet reports the same 15 pre-existing errors as
the baseline (none in the moved file or its new barrel).
